# back/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Autentication Users",
        "descriptions": "Serviço com foco em validações e gerenciamente de permissões de usuarios",
        "contact": {
            "email": "nivaldeir-silva@hotmail.com"
        }
    },
    "paths": {
        "/auth/sign-in": {
            "post": {
                "summary": "Autenticação de usuario",
                "descriptions": "Usuario pode escolher, autenticar pelo email ou password",
                "tags": ["Auth"],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "password": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "responses": {
                        "404": {
                            "description": "Not Found",
                            "content": {
                                "application/json": {
                                    "schema": {
                                        "$ref": "#/components/schemas/Error"
                                    }
                                }
                            }
                        },
                        "200": {
                            "description": "OK",
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthResponse200"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/recovery": {
            "post": {
                "summary": "Solicitação para redefinir a senha",
                "descriptions": "Solicitação para redefinir a senha",
                "tags": ["Auth"],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "responses": {
                        "404": {
                            "description": "Not Found",
                            "content": {
                                "application/json": {
                                    "schema": {
                                        "$ref": "#/components/schemas/Error"
                                    }
                                }
                            }
                        },
                        "200": {
                            "description": "OK"
                        }
                    }
                }
            }
        },
        "/auth/recovery/{token}": {
            "post": {
                "summary": "Redefinir a senha",
                "descriptions": "Redefinir a senha",
                "tags": ["Auth"],
                "parameters": [
                    {
                        "name": "token",
                        "in": "path",
                        "description": "token do usuario",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "empasswordail": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "responses": {
                        "404": {
                            "description": "Not Found",
                            "content": {
                                "application/json": {
                                    "schema": {
                                        "$ref": "#/components/schemas/Error"
                                    }
                                }
                            }
                        },
                        "200": {
                            "description": "OK"
                        }
                    }
                }
            }
        },
...
        "/permission": {
            "post": {
                "summary": "Cadastro de permissões",
                "descriptions": "",
                "tags": ["Permission"],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Permission"
                            },
                            "examples": {
                                "user": {
                                    "value": {
                                        "name": "create_user"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Já existe informação com esse...",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/Permission"
                                }
                            }
                        }
                    }
                }
            },
            "get": {
                "summary": "Busca de Permissões",
                "descriptions": "",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "tags": ["Permission"],
                "responses": {
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "$ref": "#/components/schemas/PermissionArray"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/permission/{id}": {
            "get": {
                "descriptions": "Busca de permissão pelo ID",
                "summary": "Busca de permissão pelo ID",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "tags": ["Permission"],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id do permissão",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/Permission"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "descriptions": "Atualizando um permissão pelo ID",
                "summary": "Atualizando um permissão pelo ID",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "tags": ["Permission"],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do permissao",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "examples": {
                                "user": {
                                    "value": {
                                        "name": "create_user"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/Permission"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            },
            "delete": {
                "descriptions": "Deletando uma permissão pelo ID",
                "summary": "Deletando um permissão pelo ID",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "tags": ["Permission"],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da permissão",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "PermissionArray": {
                "type": "object",
                "properties": {
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "type": "object",
                            "properties": {
                                "id": {
                                    "type": "string"
                                },
                                "name": {
                                    "type": "boolean"
                                }
                            }
                        }
                    }
                },
                "required": ["message", "data"]
            },
            "Permission": {
                "type": "object",
                "properties": {
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "type": "object",
                        "properties": {
                            "id": {
                                "type": "string"
                            },
                            "name": {
                                "type": "boolean"
                            }
                        }
                    }
                },
                "Role": {
                    "type": "object",
                    "properties": {
                        "name": {
                            "type": "string"
                        }
                    }
                },
                "Permission": {
                    "type": "object",
                    "properties": {
                        "name": {
                            "type": "string"
                        }
                    }
                },
                "Auth": {
                    "type": "object",
                    "properties": {
                        "password": {
                            "type": "string"
                        },
                        "email": {
                            "type": "string"
                        },
                        "username": {
                            "type": "string"
                        }
                    }
                },
                "ResponseError": {
                    "type": "object",
                    "properties": {
                        "error": {
                            "type": "boolean"
                        },
                        "message": {
                            "type": "string"
                        }
                    }
                }
            },
            "AuthResponse200": {
                "type": "object",
                "properties": {
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "type": "object",
                        "properties": {
                            "token": {
                                "type": "string"
                            }
                        }
                    }
                },
                "Role": {
                    "type": "object",
                    "properties": {
                        "name": {
                            "type": "string"
                        }
                    }
                },
                "Permission": {
                    "type": "object",
                    "properties": {
                        "name": {
                            "type": "string"
                        }
                    }
                },
                "Auth": {
                    "type": "object",
                    "properties": {
                        "password": {
                            "type": "string"
                        },
                        "email": {
                            "type": "string"
                        },
                        "username": {
                            "type": "string"
                        }
                    }
                },
                "ResponseError": {
                    "type": "object",
                    "properties": {
                        "error": {
                            "type": "boolean"
                        },
                        "message": {
                            "type": "string"
                        }
                    }
                }
            },
            "Error": {
                "type": "object",
                "properties": {
                    "error": {
                        "type": "boolean"
                    },
                    "message": {
                        "type": "string"
                    }
                }
            },
...
        }
    }
}